+++ tests/testthat.R
@@ -1,5 +1,6 @@
 library(testthat)
 library(tidyhydat)
 library(dplyr)
+library(lubridate)
 
 test_check("tidyhydat")