+++ tests/testthat/test_utils.R
@@ -24,6 +24,7 @@ test_that("hy_version returns a dataframe and works",{
   expect_silent(hy_version(hydat_path = hy_test_db()))
 })
 
+
 test_that("downloading hydat fails behind a proxy server with informative error message",{
   httr::set_config(httr::use_proxy(url = "http://google.com", port = 1234), override = TRUE)
   base_url <-
@@ -35,4 +36,16 @@ test_that("downloading hydat fails behind a proxy server with informative error 
                                                     "and unzipping the saved file to this directory: ",
                                                           hy_dir()))
   httr::reset_config()
+
+
+test_that("pull_station_number fails when a dataframe doesn't contain a STATION_NUMBER column",{
+  data(iris)
+  expect_error(iris %>% pull_station_number())
+})
+
+
+test_that("pull_station_number grabs station number successfully",{
+  stns <- c("08NM083", "08NE102")
+  pulled_stns <- hy_annual_stats(stns, hydat_path = hy_test_db()) %>% pull_station_number()
+  expect_identical(stns, unique(pulled_stns))
 })